Agregando BigDecimal y corrigiendo mesajes del menu

+++ src/Enteties/Moneda.java
@@ -1,5 +1,7 @@
 package Enteties;
 
-public record Moneda(String base_code, String target_code, double conversion_result) {
+import java.math.BigDecimal;
+
+public record Moneda(String base_code, String target_code, BigDecimal conversion_result) {
 
 }
